src/requirements.txt
django==1.11
djangorestframework
djangorestframework-gis
markdown
django-filter
psycopg2
requests
